refactor media caption components
